+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 Toutes les modifications notables de ce projet seront documentées dans ce fichier.
 
-## [1.2.0] - dev
+## [1.2.0] - 2025-12-18
 - choix du blockmessage, avec 403, 404, 418 ou slowfake
 - filtrage des patterns wordpress si on ne l'utilise pas
 - filtrage des pays via geoloc en liste authorisé ou non authorisé
+- Redis (en option) pour le stockage des données de session (geoloc, etc)
+- Filtrage antibot pour les bots IA (ChatGPT, Bard, etc)
 
 ## [1.1.0] - 2025-10-26
 - Controle du host lors de la redirection http vers https